Change LxWxH in Parcel to doubles

+++ EasyPost/Parcel.cs
@@ -11,9 +11,9 @@ namespace EasyPost {
         public string id { get; set; }
         public DateTime created_at { get; set; }
         public DateTime updated_at { get; set; }
-        public int length { get; set; }
-        public int width { get; set; }
-        public int height { get; set; }
+        public double length { get; set; }
+        public double width { get; set; }
+        public double height { get; set; }
         public double weight { get; set; }
         public string predefined_package { get; set; }
         public string mode { get; set; }
